+++ CHANGELOG.md
@@ -8,6 +8,12 @@ For public Changelog covering all changes done to Pipedrive’s API, webhooks an
 
 ## [Unreleased]
 
+## [28.0.0] - 2025-06-16
+### Removed
+- Removed `deals_summary` from `GET /v1/stages/:id` and `GET /v1/pipelines/:id`
+- Removed `totals_convert_currency` query parameter from `GET /v1/pipelines/:id`
+- Removed `everyone` query parameter from `GET /v1/stages/:id`
+
 ## [27.2.1] - 2025-06-09
 ### Added
 - Documented `address` property in the Organization v2 API for add and update endpoints:
@@ -900,7 +906,8 @@ Those fields will be formatted as "2020-07-13" instead of "2020-07-13T00:00:00.0
 * Fixed `GET /goal/:id/results` error handling in case when there are no existing stages connected to specified goal
 * Fixed typo in lead example response (`crrency` to `currency`)
 
-[Unreleased]: https://github.com/pipedrive/api-docs/compare/v27.2.1...HEAD
+[Unreleased]: https://github.com/pipedrive/api-docs/compare/v28.0.0...HEAD
+[28.0.0]: https://github.com/pipedrive/api-docs/compare/v27.2.1...v28.0.0
 [27.2.1]: https://github.com/pipedrive/api-docs/compare/v27.2.0...v27.2.1
 [27.2.0]: https://github.com/pipedrive/api-docs/compare/v27.1.1...v27.2.0
 [27.1.1]: https://github.com/pipedrive/api-docs/compare/v27.1.0...v27.1.1